UE 4.26.1 and C++17 support.
For UE4 versions 4.26.1 and newer use the `TInvokeResult` type trait provided by UE4.
For previous versions, check if the project is using C++17 and use `std::invoke_result_t` instead of `std::result_of_t` as it is deprecated.

MSVC might not define the correct value for `__cplusplus` depending on the version used so the `_MSVC_LANG` macro must be checked as well.